+++ src/main/java/wooteco/subway/exception/line/LineNotFoundException.java
@@ -1,4 +1,6 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.line;
+
+import wooteco.subway.exception.CustomException;
 
 public class LineNotFoundException extends CustomException {
